+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and tries to adhere (due to the nature of the project itself), to [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0.0.html).
-
+## 1.0.0 - 2022-01-01
+### Added
+- [File](TestCodeFonts.txt) to test code fonts (ligatures).
+### Changed
+- [Script](font-patcher-helper.py) code updated.
+- [Documentation files](README.md).
+- [License information](LICENSE.md).
 ## 0.0.1 - 2021-12-31
 ### Added
 - Initial script.